Moved menu related method into menu controller from rest api
